app/pyproject.toml
[tool.poetry]
name = "keiba-hacke"
version = "0.1.0"
description = ""
authors = ["nokoxxx1212"]
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
jupyterlab = "^4.0.8"
scikit-learn = "^1.4.1.post1"
lightgbm = "^4.3.0"

[tool.poetry.group.test]
optional = true
[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"

[tool.poetry.group.lint]
optional = true
[tool.poetry.group.lint.dependencies]
isort = "^5.12.0"
black = "^23.10.1"
pyproject-flake8 = "^6.1.0"

[tool.poetry.group.sample]
optional = true
[tool.poetry.group.sample.dependencies]
pandas = "2.1.2"
google-cloud-pipeline-components = "^2.9.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.kh]
optional = true
[tool.poetry.group.kh.dependencies]
pandas = "2.1.2"
google-cloud-pipeline-components = "^2.9.0"
python-dotenv = "^1.0.1"
pandas-gbq = "^0.21.0"
requests = "^2.31.0"

[tool.pytest.ini_options]
testpaths = ["tests",]